Update my wsl installer to use nvim > vim
Honestly, my original justification for this turned out to be entirely
bogus. I thought it was an issue with vim 8.1 (which was the only
trivial vim install I could get from aptget) but I think it actually was
an issue in my vimrc that i was setting the number of rows and cols in
hardcoded way as a way to set the GUI window since but not restricting
that setting to the GUI mode. Hence, whe I created a terminal instance
it would try to set the rows and columns and the wrapping would get
fucked

Ultimately, I'm keeping this because I do have some expectation that
eventually I'll have to migrate to nvim and it'll be nice to split my
time between nvim and gvim to keep comfortable with both